refactor: rename abstract to generic
